Fixed parser expressions for absorbs Changed Health/Min to Health/Sec Added Health/Sec calculation
Note: discovered a bug in recent damage/healing calculations. If the recent interval spans back before midnight, the timestamp of the earlier entries will be treated as if they were from 11:59:59pm of the current day, rather than the previous day. This will make the actual interval less than zero which will use a default interval of 1 second. (Possible fix: if the interval is actually negative, rather than less than 1, keep adding 24 hours to it until it becomes positive.)